requirements.txt
Flask==2.2.5
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==0.21.0
openai==1.13.3
